# docs/swagger.yaml
basePath: /api/v1
definitions:
  subscriptions-service_internal_models.CostResponse:
    properties:
      currency:
        type: string
      period_end:
        type: string
      period_start:
        type: string
      total_cost:
        type: integer
    type: object
  subscriptions-service_internal_models.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        minimum: 1
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  subscriptions-service_internal_models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  subscriptions-service_internal_models.ListSubscriptionsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/subscriptions-service_internal_models.SubscriptionResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  subscriptions-service_internal_models.SubscriptionResponse:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  subscriptions-service_internal_models.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        minimum: 1
        type: integer
      service_name:
        type: string
      start_date:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST-сервис для агрегации данных об онлайн-подписках.
  title: Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    get:
      parameters:
      - description: Filter by user UUID
        in: query
        name: user_id
        type: string
      - description: Filter by service name
        in: query
        name: service_name
        type: string
      - description: Page size (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptions-service_internal_models.ListSubscriptionsResponse'
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      parameters:
      - description: Subscription payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/subscriptions-service_internal_models.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/subscriptions-service_internal_models.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subscriptions-service_internal_models.ErrorResponse'
      summary: Create a subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/subscriptions-service_internal_models.ErrorResponse'
      summary: Delete a subscription
      tags:
      - subscriptions
    get:
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptions-service_internal_models.SubscriptionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/subscriptions-service_internal_models.ErrorResponse'
      summary: Get subscription by ID
      tags:
      - subscriptions
    patch:
      consumes:
      - application/json
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/subscriptions-service_internal_models.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptions-service_internal_models.SubscriptionResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/subscriptions-service_internal_models.ErrorResponse'
      summary: Update a subscription
      tags:
      - subscriptions
  /subscriptions/cost:
    get:
      parameters:
      - description: MM-YYYY
        in: query
        name: period_start
        required: true
        type: string
      - description: MM-YYYY
        in: query
        name: period_end
        required: true
        type: string
      - description: Filter by user UUID
        in: query
        name: user_id
        type: string
      - description: Filter by service name
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscriptions-service_internal_models.CostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/subscriptions-service_internal_models.ErrorResponse'
      summary: Calculate total cost for a period
      tags:
      - subscriptions
swagger: "2.0"